- moved cmr lib to separate project - new gl events:    - on_new_modelview_matrix    - on_new_projection_matrix    - on_animation_time_changed - override gl events now optional - moved dicom importer cmr to new cmr project

+++ src/bkGL/renderable/image/dvr/DVRImage4DView.cpp
@@ -126,7 +126,7 @@ namespace bk
   /// @}
 
   /// @{ -------------------------------------------------- EVENTS
-  void DVRImage4DView::on_animation_time_changed(double d)
+  void DVRImage4DView::on_animation_time_changed(GLfloat d)
   {
       const auto& image_size = _pdata->img.geometry().size();
       const auto& image_scale = _pdata->img.geometry().transformation().scale();
